requirements.txt
google-generativeai
tkhtmlview 
markdown2